backend/requirements.txt
flask==3.0.2
transformers==4.38.2
torch==2.2.1
numpy==1.26.4
gunicorn==21.2.0
python-dotenv==1.0.1
flask-cors==4.0.0
pytest==8.0.2
pytest-cov==4.1.0
flask-restx==1.3.0
openai-whisper
ffmpeg-python 